+++ src/features/bgg-search/components/game-details/game-details.component.html
