+++ src/features/bgg-search/components/game-details/game-details.component.scss
@@ -1,5 +1,5 @@
-@use "../../core/styles/variables" as variables;
-@use "../../core/styles/scrollbar.scss";
+@use "../../../../core/styles/variables" as variables;
+@use "../../../../core/styles/scrollbar.scss";
 
 .details {
   &__container {
