+++ src/features/games/components/games/games.component.ts
@@ -26,14 +26,14 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSelectChange, MatSelectModule } from '@angular/material/select';
 
-import { GameCard } from '../../../../components/models';
+import { GameCard } from '../../../../shared/models';
 import { Subject, debounceTime, distinctUntilChanged, takeUntil } from 'rxjs';
 import { FilterFunctionsService } from '../../../../core/functions/filter/filter-functions.service';
-import { HighlightTextPipe } from '../../../../core/pipes/highlight-text/highlight-text.pipe';
+import { HighlightTextPipe } from '../../../../shared/pipes/highlight-text/highlight-text.pipe';
 import { CommonFunctionsService } from '../../../../core/functions/common/common-functions.service';
 import { HttpService } from '../../../../core/services/http/http.service';
-import { ScrollToTopBtnComponent } from '../../../../components/scroll-to-top-btn/scroll-to-top-btn.component';
-import { LoaderComponent } from '../../../../components/loader/loader.component';
+import { ScrollToTopBtnComponent } from '../../../../shared/components/scroll-to-top-btn/scroll-to-top-btn.component';
+import { LoaderComponent } from '../../../../shared/components/loader/loader.component';
 import { LoaderService } from '../../../../core/services/loader/loader.service';
 import { ExportService } from '../../../../core/services/export/export.service';
 import { MatDialog } from '@angular/material/dialog';
@@ -57,7 +57,7 @@ import { GameOfTheDayComponent } from '../game-of-the-day/game-of-the-day.compon
     ScrollToTopBtnComponent,
   ],
   templateUrl: './games.component.html',
-  styleUrl: '../../../../components/common-styles.scss',
+  styleUrl: '../../../../shared/styles/common-styles.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class GamesComponent implements OnInit, AfterViewInit {
