+++ src/features/oracles/components/oracles/oracles.component.ts
@@ -14,9 +14,9 @@ import { MatButtonModule } from '@angular/material/button';
 import { CommonFunctionsService } from '../../../../core/functions/common/common-functions.service';
 import { FilterFunctionsService } from '../../../../core/functions/filter/filter-functions.service';
 import { HttpService } from '../../../../core/services/http/http.service';
-import { ScrollToTopBtnComponent } from '../../../../components/scroll-to-top-btn/scroll-to-top-btn.component';
-import { OracleCard } from '../../../../components/models';
-import { LoaderComponent } from '../../../../components/loader/loader.component';
+import { ScrollToTopBtnComponent } from '../../../../shared/components/scroll-to-top-btn/scroll-to-top-btn.component';
+import { OracleCard } from '../../../../shared/models';
+import { LoaderComponent } from '../../../../shared/components/loader/loader.component';
 import { LoaderService } from '../../../../core/services/loader/loader.service';
 import { ExportService } from '../../../../core/services/export/export.service';
 
@@ -30,7 +30,7 @@ import { ExportService } from '../../../../core/services/export/export.service';
     ScrollToTopBtnComponent,
   ],
   templateUrl: './oracles.component.html',
-  styleUrl: '../../../../components/common-styles.scss',
+  styleUrl: '../../../../shared/styles/common-styles.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class OraclesComponent implements OnInit, AfterViewInit {
