+++ src/shared/components/loader/loader.component.html
