+++ src/shared/components/loader/loader.component.scss
@@ -1,4 +1,4 @@
-@use "../../core/styles/variables" as variables;
+@use "../../../core/styles/variables" as variables;
 
 .loader {
   &__centered {
