+++ src/shared/components/loader/loader.component.spec.ts
