+++ src/shared/components/loader/loader.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectorRef, Component, OnInit, inject } from '@angular/core';
-import { LoaderService } from '../../core/services/loader/loader.service';
+import { LoaderService } from '../../../core/services/loader/loader.service';
 
 @Component({
   selector: 'app-loader',
