+++ src/shared/components/scroll-to-top-btn/scroll-to-top-btn.component.scss
