+++ src/shared/models/bgg.model.ts
