+++ src/shared/models/catalog.model.ts
