+++ src/shared/pipes/highlight-text/highlight-text.pipe.ts
