+++ src/shared/styles/common-styles.scss
@@ -1,5 +1,5 @@
-@use "../core/styles/variables" as variables;
-@use "../core/styles/scrollbar.scss";
+@use "../../core/styles/variables" as variables;
+@use "../../core/styles/scrollbar.scss";
 
 ul,
 ol,
